Resume Matcher/requirements.txt
# Resume Matcher Dependencies
# Install with: pip install -r requirements.txt

# Core Dependencies
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
numpy>=1.24.0
requests>=2.31.0
openai>=1.0.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0

# Environment Variables
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
